added smoothscroll and adjusted email styles
This commit should add jquery (via googleapi) to the page, and a small piece of js in the head tag that makes scrolling smoother.

There are also a few adjustments made to the email to draw more attention to the fact it is a link. Namely, font size, weight and line-height.